fixed import problem
grizzly.tasks.print, even though explicitly imported anywhere as a module,
replaced built-in print, which made it impossible to debug (norweigan style).
renamed it .log_message.LogMessageTask to get around that problem.

with this fixed, the problem with test_e2e_step_async_group was found and
fixed easily. when ever running manually, grizzly-loadtester would be
picked up by env `PYTHONPATH`, this variable was not set in the context
that E2E tests ran via `BehaveContextFixture`, causing a pypi version to be
installed in the virtual env. since the test depended on un-released fixes,
this caused the test to fail.

+++ tests/fixtures.py
@@ -753,6 +753,7 @@ class BehaveContextFixture:
         self._env.update({
             'PATH': f'{str(virtual_env_path)}/bin:{path}',
             'VIRTUAL_ENV': str(virtual_env_path),
+            'PYTHONPATH': environ.get('PYTHONPATH', '.'),
         })
 
         # install dependencies
@@ -770,10 +771,12 @@ class BehaveContextFixture:
         exc: Optional[BaseException],
         traceback: Optional[TracebackType],
     ) -> Literal[True]:
-        try:
-            rmtree(self.root.parent.parent, onerror=onerror)
-        except AttributeError:
-            pass
+
+        if environ.get('KEEP_FILES', None) is None:
+            try:
+                rmtree(self.root.parent.parent, onerror=onerror)
+            except AttributeError:
+                pass
 
         return True
 
@@ -882,6 +885,7 @@ class BehaveContextFixture:
             'local',
             'run',
             '--yes',
+            '--verbose',
             feature_file,
         ]
